Eliminate public thread-pool interface
